F/wep 127 richtextblock refactoring (#109)
* Create dir richTextBlock.

* Create formats.ts.

* Rename richtextblock to slate.

* Create file slateUtils.ts

* Avoid package name in filename - more generic.

* Refactor toolbar.

* Simplify.

* Generic submenu button.

* Better table structure.

* Add local context to submenu.

* Fix linting.

* Restructure dir.

* Fix.

* Update snapshots.

* Improve module structure.

* Correct translaction.

* Correct imports.

* Update snapshots.

* Correct import.
