+++ packages/editor/src/client/atoms/emojiPicker.less
